python_etl/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
pytest==8.2.2
